secure the views

+++ courses/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from .models import Course
+from django.contrib.auth.decorators import login_required
 from .forms import CourseForm
 from django.contrib import messages
 
@@ -25,8 +26,12 @@ def course_detail(request, course):
     }
     return render(request, 'courses/course_detail.html', context)
 
+@login_required
 def add_course(request):
     """add new courses to site"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admins can do that')
+        return redirect(reverse('home'))
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES)
         if form.is_valid():
@@ -45,8 +50,13 @@ def add_course(request):
     return render (request, template, context)
 
 
+@login_required
 def edit_course(request, course_id):
     """ Edit a course """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admins can do that')
+        return redirect(reverse('home'))
+        
     course = get_object_or_404(Course, pk=course_id)
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES, instance=course)
@@ -70,8 +80,13 @@ def edit_course(request, course_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_course(request, course_id):
     """ delete a course """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only site admins can do that')
+        return redirect(reverse('home'))
+
     course = get_object_or_404(Course, pk=course_id)
     course.delete()
     messages.success(request, 'Successfully deleted course!')
